Passive Sensor Part 1
The passive sensor component, as well as some of the ship data relating
to them has been finished. Armor and Engine costs were converted to
decimal to keep consistent with pre-existing ship costs, and an
additional test was put into ship tests.

The next part will be putting the rest of the ship,taskforce, and
faction level work for sensors in.

ActiveSensor is just a stub for right now.