Large Faction Part 2 and Sensor model mechanics
The sensor model stuff isn't yet tested, but it is now present.

Faction and Starsystem are now linked to a greater degree than before,
both need to know certain things about each other and there are plenty
of loose ends that need to be tied up.

Upon a faction having a presence in a new system a faction detectionList
needs to be created and propagated to the system in question. Likewise
for when that faction leaves.

Every time a contact enters or leaves a system(by whatever means) the
SystemContactList needs to be updated as appropriate.

Faction now needs to know FactionID, an integer that determines this
faction's place in the overall list of factions. This is because
starsystem now stores FactionDetectionLists, which require this
factionID to match up to the appropriate faction. Likewise ship and
population thermal,EM, and active detection markers also require
FactionID to be similar.

SensorSweep(TimeSlice) in Faction.cs is largely responsible for most of
this, and Of course I have as stated above yet to test it.