Passive sensor bug fix and ship order type mostly filled in
taskgroups can now handle not having a passive sensor suite available to
them, some component code was moved over to componentListTN, and most of
the orders are in place.

As far as orders are concerned, almost all of them are some variety of
1)go here, 2)do something on arrival, which should be simple enough to
accomplish.

More sophisticated automation orders will be done much later of course.