Components and the Message log part 1
Cargo handling for components has been started, but colonies as yet have
no place to store them, so that will need to be done. Likewise all
components have had a type of component listed, as well as a component
index for the eventual ship lists of components.

On the message log front, the MessageEntry class was created, and as
soon as timing is worked out and hammered down I'll be able to have ship
events put messages in the faction log.