Destroy and Repair component testing started.
The big challenge for repair component seems to be getting it the
necessary information, that will have to be handled via destroyed
components and the damage control que when those are done. also a more
in depth stress test will be called for once the entire damage and
targeting model is complete.