Minimal Ship to ship combat engine completed and tested
Many numerous bugs were eliminated, and the basic combat engine is in
the function SimulationTest in ShipTests.cs. alter the statistics to the
two ships to your hearts content and send them at it.

They just rush at each other, and then fire when they are able to for
now.