FactionContact logic
FactionContact is a new list of only detected contacts for each faction.
After sensorsweep is run on a taskgroup, the ships in the taskgroup have
their sensor info checked, and updated if applicable. Newly detected
ships are marked as such, and ships that are no longer detected are
removed from the list. This part of the sensor code at least is quite
simple, it has yet to be tested yet however. It should resolve
outstanding issues with the sensor model for combat simulations.