Ship to Ship Combat 0.2 finished
Support for multiple factions, and the detectedContacts dictionary are
present now, and the sim can handle destroyed ships mid simulation.

16 Ships/taskgroups are randomly created and scattered about the test
system, and their orders are to either follow a target and destroy it,
or head for the center of the test system.

Also there is a stub for wreck, not doing anything with that yet, and I
suspect for a long time.