Ship To Ship 0.3
1280 ships are spawned randomly around a 2 AU area, and sent at each
other. They are divided up into 16 factions, 10 TGs per faction, 8 ships
per TG. it takes between 30 minutes and an hour of SimTime(not run time)
for the battle to finish as is.

Additional features at this point are a more sophisticated AI, or other
TN game mechanics, so I think this particular task is finished.