Taskgroup page Minimal functionality
Now the taskgroup page can set speed and destination for taskgroups.
Right now contacts(not yet tested), other taskgroups, waypoints and
planets can all be destination targets.

Names were also added to several classes in conjunction with these
changes. Waypoints are not yet faction specific, that will have to be
changed. Also a bug in SimEntity was fixed, it will not create multiple
copies of the same first faction now.