Ship design UI with respect to missiles is almost complete
Missiles are designable, can be added to ship designs, subtracted from
ship designs, the display will check to see if the missile is obsolete
or larger than the largest launch tube on the vessel, and all conditions
seem handled appropriately.

Only the load ships button remains to be done for missiles in class
design, and that will be done when I implement missile combat.