Magazine and Launch Tube destruction now handled
Ships with no magazines have an ammo limit of MaxLauncherCount, this is
reflected in both the ability to load missiles to the preferred ordnance
list, and the ammunition display in taskgroup.

Fixed a display bug for size tonnage.

New member variables tracking each component of Magazine space(whether
from launch tubes or magazines) added.

Individual Magazine handling for magazine ordnance removed, this is no
longer how missiles will be tracked.

And to the meat of the update: Missiles are considered load balanced
across all magazines, and this is reflected in magazine destruction.
Likewise ordnance launch tubes will be considered loaded if there is an
assigned ordnance OR there is no assigned ordnance, but not enough
magazine specific space exists to support all missiles.