Missile Combat part 1:
Provisional and initial missile combat commit:
Missile groups will be created, and will go to their target(and do
damage when they get there). Many conditions are not properly handled
yet, and there is no missile icon as yet either, so they will not be
drawn, but the rest of the UI indicates that missiles were launched and
damage was done. Point defense is also not yet possible.

Missile ships have to be created and fastoobed in, I will probably add a
missile ship to both sides when finished.