TurretDefTN done.
This is the definition for a turret, the actual work of connecting it to
the back and front end, and then handling its mechanics is not yet
complete. TurretDefTN will, in addition to requiring everything CIWS
required, require changes to beam fire control as it is a new weapon
type that can be linked to a turret. Currently there is only a single
binding list for BeamTN, an additional one must be made for TurretTN and
likewise the UI must be updated to reflect this.