CIWS implementation done
This is not yet tested, but CIWS should be able to intercept missiles
now. Likewise a system has been worked out for how to keep track of
which CIWS have fired, since they do not have the reactor requirements
that other beams have, and will want to not expend all shots on the
first missile they come across.

next up, Final Defensive Fire(Self), which should be largely the same as
the CIWS implementation. FDF regular will be somewhat more tricky, since
a distance check of PD enabled FCs will be required. Area defense will
also require this but will be handled in SimEntity rather than on
missile impact.