Final defensive fire self done.
PD is still not tested, but the implementation of FDF_Self is laid out
here.

Additionally, a new order type constructor for contacts was created, new
logic will have to be put in to use it however. this will be necessary
as contacts are tracked on a ship by ship basis, but orders to contacts
are handled at a taskgroup level, and this can lead to issues with
contacts disappearing or being destroyed that are not presently handled.