Area Defense Part 1
This is a rough draft, make it work layout for how Area defense will
work. First the Fire Controls will be looped through until an
appropriate one is found, then all the missiles in the star system that
FC is in have to be tested. The actual work of shooting down missiles,
and marking them as destroyed/removing ordnance groups from simulations
is not done for area defense yet.

Later on separating the PD FC lists out by star system will make this
more efficient if that is required.