Economics page tree view and time control functionality
The Tree view for the economics page should be totally functional now,
though it will have to be tested, and time control works as well as the
existing system map time controls do, complete with map updates when the
economics page time controls are used.

Later the system map should update the economics page, when the econ
page is implemented.

Also basic ruins code is included.