Shipyard Partial Work Part 1
Shipyards are now buildable, and some tasks can be assigned to them.
There is no construction cycle for shipyards as yet however.

An issue with partial construction factory construction was resolved,
and now items are charged on the basis of their amount of completion
each construction cycle, not in a lump sum up front anymore.

Ordnance Targeting populations and bodies has had the primary adjustment
removed to reflect that this should no longer be necessary. be aware of
that in case it turns out that I am wrong here.

In the future I will try to put all economics work into secondary helper
files like Eco_ShipyardTabHandler, and will try to refactor the existing
code to cut down on the size of economics.cs.