Recommitting this in case I missed it
The goal with the floating point adjustment here is to give a more
random band for squadron transits to jump into, I'm not sure how exactly
they are handled in Aurora, but this seems reasonable. 5% of jump radius
is fairly tight and still allows for nearly full use of the jumpdrive's
capability while still increasing the area the ship can appear in. Still
if there is a better way to do this feel free to change it.