Shipyard backend work
This allows for Shipyard complexes to perform activities as well as
tasks(adding tonnage and building new ships for example), but is not
tested. Likewise the UI still has a little bit further to go as tasks
are not displayed yet. Finally there are still a few extreme conditions
regarding what to do if the shipyard or ship in the shipyard gets
destroyed before task completion.