minor issues roundup #1
TGs are created if there is no Taskgroup available to add a ship to at a
shipyard, and the UI elements for adding or removing ships from
taskgroups have been added. Still need to do the back end work for ship
changes between taskgroups and  SYTaskGroupComboBox may have to get
looked at if a new TG is created by the above work.