Survey displayfix and addition to TG locations
survey points are now displayed properly on all zoom levels, and can be
added as a location for a taskgroup to target. The actual survey
mechanics are not yet implemented, and will require storage within a
starsystem of the results.  I'll want to have a bool for a totally
completed survey of a starsystem, as well as some way to store partial
results since it will be unlikely for any player to leave a starsystem
in an incompletely surveyed state, and thus having to constantly iterate
through 30 points to determine if a system is fully surveyed will be
unnecessary most of the time.