Survey in progress work
Each system now keeps a list of what surveying is currently in progress,
and taskgroups will now check on order removal to make sure that this
list does not have spurious entries in it. this is important for default
orders to work properly, and make sure that multiple ships don't
re-survey the same point or body.