Additional fixes
The shipyard tonnage bug was fixed, I was setting the tonnage to 1000
after checking for yard specific tonnage. woops.

Shipyards can be set to No Activity now, this was an oversight with
setActivity that it could not.

Geosurvey distances were being calculated incorrectly, they are now
fixed.

The treeview is now properly selectable, key and mouse press events were
happening before the selection was being made, switching to afterSelect
resolves this.

The agricultural and environmental percentages are spurious and will
need to be looked at, and any issue from last update that was not
resolved here is likely still a problem