Several more bugfixes
duplicate populations were showing up in various places, that has been
fixed(this caused issues with the tree view having multiples of the same
item added to dictionaries).
Gravitational Survey Sensors are now military components, before all
survey sensors were civilian.
The cost listbox for the currently selected item to build on the
economics page is now being refreshed on a softRefresh(time advancement,
which does not need a full refresh of the entire page).

Designs don't seem to be locking immediately, I'm not sure how to
resolve this, it isn't immediately important however. Shipyard eligible
classes require all such designs to be locked in order to function.
The CI rounding bug is still present, and as with the above, I'm not
quite sure what to do about it.