Jump orders fix and others
when jumping to a new system, CanOrder was not properly being set to
AcceptOrders, this is fixed. likewise the homeworld mineral generation
kludge was removed, and now body mineral generation is handled in
systembody.cs for the most part.

Construction factories were costing Corrundum instead of vendarite, that
is fixed for when that all gets moved to a json file.

Ships failing 1 conditional order do not warn if another conditional
succeeds now, and warnings for no geosurvey target are now correctly
labeled as geosurvey target failures.