Load and unload taskgroup instruction work
loads and unloads for minerals should now work, with both the ships
cargo lists, as well as previous orders being looked at to figure out
what to load and unload. It is currently up to the user to not mess up
unload orders, if they do it will crash.

Also a sensor bug was fixed, and a minor issue with ship naming for
elligible classes may have been resolved.