UI and Colony bugfixes
Colonists and Infrastructure can now be moved from one colony to
another, but there are no repeating orders just yet. Also the annual
growth rate for colonies is not negative if they exceed their
infrastructure.

Other minor issues to look at are that costs for cap expansion for
shipyards may be off, and double clicking on an item in the secondary
orders listbox should issue an order.